Extract `assert_output` and `available_pty?` into `ConsoleHelpers` module
We define almost the same method with multiple tests. Therefore, it extract
into module.
